Refactor Borrow logic
