+++ src/pages/nftLending/BorrowPage/InstantLoansContent/components/BorrowTable/hooks.ts
@@ -27,12 +27,8 @@ import {
 
 import { useCartState } from '../../cartState'
 import { getTableColumns } from './columns'
-import {
-  createTableNftData,
-  executeBorrow,
-  makeCreateTxnsDataParams,
-  showBonkRewardsSnack,
-} from './helpers'
+import { executeBorrow, makeCreateTxnsDataParams } from './core'
+import { createTableNftData, showBonkRewardsSnack } from './helpers'
 import { TableNftData } from './types'
 
 import styles from './BorrowTable.module.less'
