+++ src/pages/nftLending/BorrowPage/InstantLoansContent/components/BorrowTable/index.ts
@@ -1,1 +1,3 @@
 export { default } from './BorrowTable'
+export * from './helpers'
+export * from './core'
